+++ src/structures/Interaction.js
@@ -1,8 +1,8 @@
 'use strict';
 
+const { InteractionType } = require('discord-api-types/v9');
 const { Base, Permissions } = require('discord.js');
 const SnowflakeUtil = require('../util/SnowflakeUtil');
-const { InteractionType } = require('discord-api-types/v9');
 
 /**
  * Represents an Interaction.
@@ -15,7 +15,7 @@ class Interaction extends Base {
 
     this.id = data.id;
 
-    Object.defineProperty(this, "token", { value: data.token });
+    Object.defineProperty(this, 'token', { value: data.token });
 
     this.applicationId = data.application_id;
 
@@ -25,15 +25,11 @@ class Interaction extends Base {
 
     this.user = this.client.users._add(data.user ?? data.member.user);
 
-    this.member = data.member
-      ? this.guild?.members._add(data.member) ?? data.member
-      : null;
+    this.member = data.member ? this.guild?.members._add(data.member) ?? data.member : null;
 
     this.version = data.version;
 
-    this.memberPermissions = data.member?.permissions
-      ? new Permissions(data.member.permissions).freeze()
-      : null;
+    this.memberPermissions = data.member?.permissions ? new Permissions(data.member.permissions).freeze() : null;
 
     this.locale = data.locale;