+++ src/structures/ModalSubmitInteraction.js
@@ -1,11 +1,12 @@
-const { InteractionWebhook } = require("discord.js");
-const ModalSubmitField = require("./ModalSubmitField");
-const ModalActionRow = require("./ModalActionRow");
-const BaseMessageComponent = require("./BaseMessageComponent");
-const Interaction = require("./Interaction");
-const InteractionResponses = require("./interfaces/InteractionResponses");
-const { InteractionTypes } = require("../util/Constants");
-const { Message } = require("discord.js");
+'use strict';
+
+const { InteractionWebhook } = require('discord.js');
+const BaseMessageComponent = require('./BaseMessageComponent');
+const Interaction = require('./Interaction');
+const ModalActionRow = require('./ModalActionRow');
+const ModalSubmitField = require('./ModalSubmitField');
+const InteractionResponses = require('./interfaces/InteractionResponses');
+const { InteractionTypes } = require('../util/Constants');
 
 /**
  * Represents a Modal Submit Interaction.
@@ -21,8 +22,8 @@ class ModalSubmitInteraction extends Interaction {
      * @type {String}
      */
 
-    this.type = InteractionTypes[data.type] ?? null;   
-    
+    this.type = InteractionTypes[data.type] ?? null;
+
     /**
      * The Custom Id of the Modal.
      * @type {String}
@@ -37,7 +38,7 @@ class ModalSubmitInteraction extends Interaction {
 
     this.message = data.message ? this.channel?.messages._add(data.message) ?? data.message : null;
 
-    let modalFields = [];
+    const modalFields = [];
     for (let i = 0; i < data.data.components.length; i++) {
       const field = data.data.components[i].components[0];
       modalFields.push(field);
@@ -48,27 +49,25 @@ class ModalSubmitInteraction extends Interaction {
      * @type {ModalSubmitField}
      */
 
-    this.fields = modalFields.map((f) => new ModalSubmitField(f)) ?? data.data.components?.map((c) => BaseMessageComponent.create(c, this.client)) ?? [];
+    this.fields =
+      modalFields.map(f => new ModalSubmitField(f)) ??
+      data.data.components?.map(c => BaseMessageComponent.create(c, this.client)) ??
+      [];
 
     /**
      * The Action Rows of the modal with the Text Input Components.
      */
 
-    this.components = data.data.components?.map((component) => new ModalActionRow(component))
+    this.components = data.data.components?.map(component => new ModalActionRow(component));
 
     /**
      * An associated interaction webhook, can be used to further interact with this interaction
      * @type {InteractionWebhook}
      */
 
-    this.webhook = new InteractionWebhook(
-      this.client,
-      this.applicationId,
-      this.token
-    );
+    this.webhook = new InteractionWebhook(this.client, this.applicationId, this.token);
   }
 
-  
   /**
    * Gets a Text Input Component value.
    * @param {string} customId The Custom Id of a Text Input Component.
@@ -76,9 +75,9 @@ class ModalSubmitInteraction extends Interaction {
    */
 
   getTextInputValue(customId) {
-    const field = this.fields.find((field) => field.customId === customId);
-    
-    return field ? field.value : null;
+    const fieldFound = this.fields.find(f => f.customId === customId);
+
+    return fieldFound ? fieldFound.value : null;
   }
 
   /**
@@ -88,13 +87,13 @@ class ModalSubmitInteraction extends Interaction {
    */
 
   getField(customId) {
-    const field = this.fields.find((field) => field.customId === customId);
+    const fieldToGet = this.fields.find(f => f.customId === customId);
 
-    return field ? field : null;
+    return fieldToGet ? fieldToGet : null;
   }
 
   isFromMessage() {
-    return Boolean(this.message)
+    return Boolean(this.message);
   }
 
   deferReply() {}
@@ -106,9 +105,6 @@ class ModalSubmitInteraction extends Interaction {
   update() {}
 }
 
-InteractionResponses.applyToClass(ModalSubmitInteraction, [
-  "deferUpdate",
-  "showModal",
-]);
+InteractionResponses.applyToClass(ModalSubmitInteraction, ['deferUpdate', 'showModal']);
 
 module.exports = ModalSubmitInteraction;