+++ src/util/Constants.js
@@ -1,3 +1,5 @@
+'use strict';
+
 exports.TextInputStyles = createEnum([null, 'SHORT', 'LONG']);
 
 exports.MessageComponentTypes = createEnum([null, 'ACTION_ROW', 'BUTTON', 'SELECT_MENU', 'TEXT_INPUT']);
@@ -25,11 +27,11 @@ exports.InteractionTypes = createEnum([
 ]);
 
 function createEnum(keys) {
-    const obj = {};
-    for (const [index, key] of keys.entries()) {
-      if (key === null) continue;
-      obj[key] = index;
-      obj[index] = key;
-    }
-    return obj;
+  const obj = {};
+  for (const [index, key] of keys.entries()) {
+    if (key === null) continue;
+    obj[key] = index;
+    obj[index] = key;
   }
+  return obj;
+}